ADD library app. MODIFIED books models ( move model from users to library)
