+++ frontend/templates/_base.html
@@ -37,7 +37,7 @@
                   <a class="nav-link" href="{% url 'home' %}"><i class="fas fa-home fa-3x"></i></a>
                 </li>
                 <li class="nav-item">
-                  <a class="nav-link" href="#" ><i class="fas fa-book fa-3x"></i></a>
+                  <a class="nav-link" href="{% url 'library'%}" ><i class="fas fa-book fa-3x"></i></a>
                 </li>
                 <li class="nav-item">
                   <a class="nav-link" href="#"><i class="fas fa-book-medical fa-3x"></i></a>
@@ -46,7 +46,7 @@
                   <a class="nav-link dropdown-toggle" href="#" id="dropdownId" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fas fa-portrait fa-3x"></i></a>
                   <div class="dropdown-menu" aria-labelledby="dropdownId">
                     {% if user.is_authenticated %}  
-                    <a class="dropdown-item" href="#">My Account </a>
+                    <a class="dropdown-item" href="{% url 'mybooks' %}">My Account </a>
                     <a class="dropdown-item" href="{% url 'account_logout' %}">Log Out </a>
                     {% else%}                
                     <a class="dropdown-item" href="{% url 'account_login' %}">Log In </a>
