+++ frontend/templates/library/mybooks.html
@@ -1,0 +1,9 @@
+{% extends '_base.html' %}
+{% block title %}Books{% endblock title %}
+{% block content %}
+{% for book in mybooks %}
+<div>
+    <h2><a href="{{ book.get_absolute_url }}">{{ book.title }}</a></h2>
+</div>
+{% endfor %}
+{% endblock content %}
