--- frontend/templates/mybooks/book_detail.html
@@ -1,8 +1,0 @@
-{% extends '_base.html' %}
-{% block title %}{{ book_detail.title }}{% endblock title %}
-{% block content %}
-<div class="book-detail">
-    <h2><a href="">{{ book_detail.title }}</a></h2>
-    <p>Author: {{ book_detail.author }}</p>
-</div>
-{% endblock content %}
