--- frontend/templates/mybooks/books_list.html
@@ -1,9 +1,0 @@
-{% extends '_base.html' %}
-{% block title %}Books{% endblock title %}
-{% block content %}
-{% for book in books_list %}
-<div>
-    <h2><a href="{{ book.get_absolute_url }}">{{ book.title }}</a></h2>
-</div>
-{% endfor %}
-{% endblock content %}
