+++ users/views.py
@@ -1,13 +1,9 @@
 
 from django.shortcuts import render
 from django.views import generic
-from django.views.generic import ListView, DetailView
 from django.views.generic.edit import CreateView
 from django.urls import reverse_lazy
 from .forms import CustomUserCreationForm
-from .models import MyBook
-
-
 
 
 class SignUpPageView(generic.CreateView):
@@ -16,24 +12,6 @@ class SignUpPageView(generic.CreateView):
     template_name = 'signup.html'
 
 
-class BookListView(ListView):
-           
-    model = MyBook
-    context_object_name = 'books_list'
-    template_name = 'mybooks/books_list.html'  
 
-    def get_queryset(self):
-        # original qs
-        qs = super().get_queryset()
-        # filter by a variable captured from url, for example
-        return qs.filter(owner=self.request.user)
-    
-        
-    
-
-class BookDetailView(DetailView):
-    model = MyBook
-    context_object_name = 'book_detail' 
-    template_name = 'mybooks/book_detail.html'
